fixed readme typos, improved readbility,created lib folder for logic files, installed colorize gem, fixed terminal readability
