+++ Gemfile
@@ -6,5 +6,6 @@ source "https://rubygems.org" do
     gem "rubocop"
     gem "bundler"
     gem 'rspec'
+    gem 'colorize'
 
     end
